requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.5

# Authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
asyncpg>=0.28.0

# Google Cloud
google-cloud-logging>=3.5.0
google-cloud-storage>=2.10.0

# AI/ML
google-generativeai>=0.3.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
openai>=1.6.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.1

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.1.0
httpx>=0.24.1
sse-starlette>=1.6.5

# Redis
redis>=5.0.0